Add TestRenderer, fixup test
